Organise commands, parser, storage, ui into their own packages
