+++ src/main/java/duke/commands/CreateDeadlineCommand.java
@@ -1,0 +1,9 @@
+package duke.commands;
+
+import duke.task.Deadline;
+
+public class CreateDeadlineCommand extends CreateTaskCommand {
+    public CreateDeadlineCommand(Deadline deadline) {
+        super(deadline);
+    }
+}
