+++ src/main/java/duke/commands/CreateEventCommand.java
@@ -1,0 +1,9 @@
+package duke.commands;
+
+import duke.task.Event;
+
+public class CreateEventCommand extends CreateTaskCommand {
+    public CreateEventCommand(Event event) {
+        super(event);
+    }
+}
