+++ src/main/java/duke/commands/CreateTodoCommand.java
@@ -1,0 +1,9 @@
+package duke.commands;
+
+import duke.task.Todo;
+
+public class CreateTodoCommand extends CreateTaskCommand {
+    public CreateTodoCommand(Todo todo) {
+        super(todo);
+    }
+}
