+++ src/main/java/duke/commands/ExitCommand.java
@@ -1,0 +1,12 @@
+package duke.commands;
+
+import duke.storage.Storage;
+import duke.task.TaskList;
+import duke.ui.Ui;
+
+public class ExitCommand extends Command {
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) {
+        ui.showGoodByeMessage();
+    }
+}
