+++ src/main/java/duke/exceptions/InvalidDeadlineException.java
@@ -1,4 +1,4 @@
-package duke.Exceptions;
+package duke.exceptions;
 
 public class InvalidDeadlineException extends InvalidCommandException {
     public InvalidDeadlineException() {
