+++ src/main/java/duke/task/Deadline.java
@@ -1,5 +1,5 @@
-package duke.Task;
-import duke.Exceptions.*;
+package duke.task;
+import duke.exceptions.*;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
