+++ src/main/java/duke/task/Todo.java
@@ -1,5 +1,5 @@
-package duke.Task;
-import duke.Exceptions.*;
+package duke.task;
+import duke.exceptions.*;
 
 /**
  * Todos are tasks without any date/time attached to it e.g., visit new theme park
