+++ src/main/java/duke/ui/Ui.java
@@ -1,8 +1,8 @@
-package duke;
+package duke.ui;
 
 public class Ui {
-    private static String WELCOME_MESSAGE =  "Hello! I'm duke.Duke\n" + "What can I do for you?";
-    private static String GOODBYE_MESSAGE =  "Bye. Hope to see you again soon!";
+    private static String WELCOME_MESSAGE = "Hello! I'm duke.Duke\n" + "What can I do for you?";
+    private static String GOODBYE_MESSAGE = "Bye. Hope to see you again soon!";
 
     public void showWelcomeMessage() {
         System.out.println(WELCOME_MESSAGE);
